refactor: 함수명 통일 및 코드 일관성 개선
